requirements (1).txt
# ── Deep learning ──────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── Data / ML ──────────────────────────────────────────────────────────────────
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0

# ── Web app ─────────────────────────────────────────────────────────────────────
flask>=3.0.0
werkzeug>=3.0.0
gunicorn>=21.0.0         # production WSGI server

# ── Dataset download ───────────────────────────────────────────────────────────
kaggle>=1.6.0

# ── Optional (Grad-CAM visualisation) ─────────────────────────────────────────
# opencv-python>=4.8.0   # uncomment to enable GradCAM overlays

# ── Utilities ──────────────────────────────────────────────────────────────────
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.13.0
